requirements.txt
numpy
pandas
tensorflow-cpu
keras
Pillow